XAADMINISTRATION-77: Improve the registration interface - Run Selenium tests with LiveValidation and without, also test configuration from adminstration page and improved performance of RegisterTest
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/enterprise/trunk@27367 f329d543-caf0-0310-9063-dda96c69346f

+++ distribution-test/selenium-tests/src/test/it/com/xpn/xwiki/it/selenium/RegisterTest.java
@@ -43,24 +43,39 @@ public class RegisterTest extends AbstractXWikiTestCase
     {
         super.setUp();
 
+        // We start at localhost:4444/selenium-server... and isExistingPage reloads current location 
+        // so navigate to main page.
+        if (!isAuthenticationMenuPresent()) {
+            open("Main", "WebHome");
+        }
+
         // Remove "JohnSmith" user if already exists
-        loginAsAdmin();
-        deletePage("XWiki", "JohnSmith");
+        if (isExistingPage("XWiki", "JohnSmith")) {
+            loginAsAdmin();
+            deletePage("XWiki", "JohnSmith");
+            logout();
+        }
 
         // Ensure that the user isn't logged in
-        logout();
+        if (isAuthenticated()) {
+            logout();
+        }
 
-        clickRegister();
+        // Usually we are already at the registration page, this makes it more fragile but quicker.
+        if (!getSelenium().isTextPresent("Welcome to the registration form. This will allow you to edit pages,")) {
+            clickRegister();
+        }
+
         fillFormWithJohnSmithValues();
     }
 
-    private void fillFormWithJohnSmithValues()
-    {
-        fillRegisterForm("John", "Smith", "JohnSmith", "JohnSmith", "JohnSmith@example.com");
-    }
 
+    /* The Tests */
+
+    /* LiveValidation should have no influence on this test so it's only run once (with LiveValidation on) */
     public void testRegisterJohnSmith()
     {
+        setUseLiveValidation(true);
         submit();
         assertTextPresent("Registration successful");
         clickLinkWithLocator("link=John Smith");
@@ -70,35 +85,128 @@ public class RegisterTest extends AbstractXWikiTestCase
         login("JohnSmith", "JohnSmith", false);
     }
 
+    public void testRegisterWithoutEnteringPasswordAndPasswordTooShortWithLiveValidation()
+    {
+        setUseLiveValidation(true);
+        tryRegisterWithoutEnteringPassword(false);
+    }
+
+    public void testRegisterWithTwoDifferentPasswordsWithLiveValidation()
+    {
+        setUseLiveValidation(true);
+        tryRegisterWithTwoDifferentPasswords(false);
+    }
+
+    public void testRegisterWithoutEnteringUserNameWithLiveValidation()
+    {
+        setUseLiveValidation(true);
+        tryRegisterWithoutEnteringUserName(false);
+    }
+
+    /* LiveValidation should have no influence on this test so it's only run once (with LiveValidation on) */
     public void testRegisterExistingUser()
     {
+        setUseLiveValidation(true);
         setFieldValue("xwikiname", "Admin");
         submit();
         assertTextPresent("User already exists.");
     }
 
-    public void testRegisterWithTwoDifferentPasswords()
+    public void testSwitchLiveValidationOff()
+    {
+        setUseLiveValidation(true);
+        assertElementPresent("xpath=/html/body/div/div/div[3]/div/div/div/div/div/script");
+        setUseLiveValidation(false);
+        assertElementNotPresent("xpath=/html/body/div/div/div[3]/div/div/div/div/div/script");
+    }
+
+    public void testRegisterWithoutEnteringPasswordNoLiveValidation()
+    {
+        setUseLiveValidation(false);
+        tryRegisterWithoutEnteringPassword(true);
+    }
+
+    public void testRegisterWithTwoDifferentPasswordsNoLiveValidation()
+    {
+        setUseLiveValidation(false);
+        tryRegisterWithTwoDifferentPasswords(true);
+    }
+
+    public void testRegisterWithoutEnteringUserNameNoLiveValidation()
+    {
+        setUseLiveValidation(false);
+        tryRegisterWithoutEnteringUserName(true);
+    }
+
+    /* Helper methods */
+
+    private void fillFormWithJohnSmithValues()
+    {
+        fillRegisterForm("John", "Smith", "JohnSmith", "JohnSmith", "JohnSmith@example.com");
+    }
+
+    /* Assumes we are at the registration page and leaves us there. */
+    public void setUseLiveValidation(boolean useLiveValidation)
+    {
+        if (useLiveValidation != isElementPresent("xpath=/html/body/div/div/div[3]/div/div/div/div/div/script")) {       
+            loginAsAdmin();
+            open("XWiki", "XWikiPreferences", "admin", "section=Registration&editor=globaladmin");
+            if (useLiveValidation == true) {
+                getSelenium().check("//input[@name='XWiki.Registration_1_liveValidation_enabled'][@type='checkbox']");
+            } else {
+                getSelenium().uncheck("//input[@name='XWiki.Registration_1_liveValidation_enabled'][@type='checkbox']");
+            }
+            // Copied from AlbatrossSkinExecutor.clickEditSaveAndContinue() 
+            // because we have to click a specific saveAndContinue button.
+            submit("//input[@value='Save XWiki.RegistrationConfig'][@name='action_saveandcontinue']", false);
+            waitForCondition("(window.document.getElementsByClassName('xnotification-done')[0] != null "
+                             + "&& window.document.getElementsByClassName('xnotification-done')[0].innerHTML == 'Saved')");
+            // Ensure that the user isn't logged in
+            logout();
+            clickRegister();
+            fillFormWithJohnSmithValues();
+        }
+    }
+
+    public void tryRegisterWithTwoDifferentPasswords(boolean submit)
     {
         setFieldValue("register_password", "a");
         setFieldValue("register2_password", "b");
-        clickLinkWithXPath("//input[@type='submit']", false);
+        if (submit) {
+            clickLinkWithXPath("//input[@type='submit']", false);
+        } else {
+            // Blur event to make livevalidation react.
+            getSelenium().fireEvent("//div/dl/dd/input[@id='register_password']", "blur");
+            getSelenium().fireEvent("//div/dl/dd/input[@id='register2_password']", "blur");
+        }
         waitForElement("xpath=//form[@id='register']/div/dl/dd[5]/span");
+        waitForTextPresent("xpath=//form[@id='register']/div/dl/dd[4]/span", "Please use a longer password.");
         waitForTextPresent("xpath=//form[@id='register']/div/dl/dd[5]/span", "Your passwords aren't the same.");
     }
 
-    public void testRegisterWithoutEnteringUserName()
+    public void tryRegisterWithoutEnteringUserName(boolean submit)
     {
         setFieldValue("xwikiname", "");
-        clickLinkWithXPath("//input[@type='submit']", false);
+        if (submit) {
+            clickLinkWithXPath("//input[@type='submit']", false);
+        } else {
+            // Blur event to make livevalidation react.
+            getSelenium().fireEvent("//div/dl/dd/input[@id='xwikiname']", "blur");
+        }
         waitForElement("xpath=//form[@id='register']/div/dl/dd[3]/span");
         waitForTextPresent("xpath=//form[@id='register']/div/dl/dd[3]/span", "This field is mandatory.");
     }
 
-    public void testRegisterWithoutEnteringPassword()
+    public void tryRegisterWithoutEnteringPassword(boolean submit)
     {
         setFieldValue("register_password", "");
         setFieldValue("register2_password", "");
-        clickLinkWithXPath("//input[@type='submit']", false);
+        if (submit) {
+            clickLinkWithXPath("//input[@type='submit']", false);
+        } else {
+            // Blur event to make livevalidation react.
+            getSelenium().fireEvent("//div/dl/dd/input[@id='register_password']", "blur");
+        }
         waitForElement("xpath=//form[@id='register']/div/dl/dd[4]/span");
         waitForTextPresent("xpath=//form[@id='register']/div/dl/dd[4]/span", "This field is mandatory.");
     }
